+++ src/apps/kwenta/optimism/kwenta.perp-v1-cross-margin.contract-position-fetcher.ts
@@ -23,7 +23,7 @@ const getCrossMarginAccountsQuery = gql`
 @PositionTemplate()
 export class OptimismKwentaPerpV1CrossMarginContractPositionFetcher extends OptimismSynthetixPerpV1ContractPositionFetcher {
   groupLabel = 'PerpV1 cross-margin';
-  extraLabel = ' (v1 cross-margin)';
+  useCustomMarketLogos = true;
 
   async getAccountAddress(address: string): Promise<string> {
     const crossMarginAccountsFromSubgraph = await gqlFetch<GetCrossMarginAccounts>({