+++ src/apps/synthetix/optimism/synthetix.perp-v1.contract-position-fetcher.ts
@@ -5,9 +5,8 @@ import { OptimismSynthetixPerpContractPositionFetcher } from './synthetix.perp.c
 @PositionTemplate()
 export class OptimismSynthetixPerpV1ContractPositionFetcher extends OptimismSynthetixPerpContractPositionFetcher {
   groupLabel = 'PerpV1';
-  extraLabel = '(v1)';
 
-  marketFilter(market) {
-    return !this.isV2Market(market);
+  marketFilter(marketKey: string) {
+    return !this.isV2Market(marketKey);
   }
 }